Task level load balanced has been optimized to minimize the shard movements.

+++ storm-core/src/jvm/backtype/storm/elasticity/utils/FirstFitDoubleDecreasing.java
@@ -33,7 +33,7 @@ public class FirstFitDoubleDecreasing extends AbstractBinPacking {
         bins.clear();
         ballToBinMapping.clear();
         for(int i=0; i<nbins; i++) {
-            bins.add(new Bin(Integer.MAX_VALUE));
+            bins.add(new Bin(Long.MAX_VALUE));
             bins.get(i).index = i;
         }
